+++ docs/core/tracing.md
@@ -139,7 +139,7 @@ context for an operation using any native object.
 
     You can add annotations using `AddAnnotation()` method from Tracing
     ```c# hl_lines="9"
-    using Amazon.Lambda.PowerTools.Tracing;
+    using AWS.Lambda.PowerTools.Tracing;
 
     public class Function
     {
@@ -156,7 +156,7 @@ context for an operation using any native object.
 
     You can add metadata using `AddMetadata()` method from Tracing
     ```c# hl_lines="9"
-    using Amazon.Lambda.PowerTools.Tracing;
+    using AWS.Lambda.PowerTools.Tracing;
 
     public class Function
     {
@@ -177,7 +177,7 @@ under a subsegment, or you are doing multithreaded programming. Refer examples b
 === "Functional Api"
 
     ```c# hl_lines="8 9 10 12 13 14"
-    using Amazon.Lambda.PowerTools.Tracing;
+    using AWS.Lambda.PowerTools.Tracing;
     
     public class Function
     {
@@ -198,7 +198,7 @@ under a subsegment, or you are doing multithreaded programming. Refer examples b
 === "Multi Threaded Programming"
 
     ```c# hl_lines="13-16"
-    using Amazon.Lambda.PowerTools.Tracing;
+    using AWS.Lambda.PowerTools.Tracing;
 
     public class Function
     {
