+++ libraries/src/AWS.Lambda.PowerTools.Common/InternalsVisibleTo.cs
@@ -15,7 +15,7 @@
 
 using System.Runtime.CompilerServices;
 
-[assembly: InternalsVisibleTo("AWS.Lambda.PowerTools.Tests")]
+[assembly: InternalsVisibleTo("AWS.Lambda.PowerTools.Common.Tests")]
 [assembly: InternalsVisibleTo("AWS.Lambda.PowerTools.Tracing.Tests")]
 [assembly: InternalsVisibleTo("AWS.Lambda.PowerTools.Metrics.Tests")]
 [assembly: InternalsVisibleTo("AWS.Lambda.PowerTools.Logging.Tests")]
