+++ libraries/src/AWS.Lambda.PowerTools.Logging/AWS.Lambda.PowerTools.Logging.csproj
@@ -20,11 +20,11 @@
     </PropertyGroup>
       
     <ItemGroup>
-      <ProjectReference Include="..\AWS.Lambda.PowerTools\AWS.Lambda.PowerTools.csproj" />
+      <ProjectReference Include="..\AWS.Lambda.PowerTools.Common\AWS.Lambda.PowerTools.Common.csproj" />
     </ItemGroup>
     
     <ItemGroup>
-        <None Include="README.md" Pack="true" PackagePath="\"/>
+        <None Include="README.md" Pack="true" PackagePath="\" />
     </ItemGroup>
 
     <ItemGroup>
