+++ libraries/src/AWS.Lambda.PowerTools.Logging/Internal/PowerToolsConfigurations.cs
@@ -14,7 +14,7 @@
  */
 
 using System;
-using AWS.Lambda.PowerTools.Core;
+using AWS.Lambda.PowerTools.Common;
 using Microsoft.Extensions.Logging;
 
 namespace AWS.Lambda.PowerTools.Logging.Internal;
