+++ libraries/src/AWS.Lambda.PowerTools.Logging/LoggingAttribute.cs
@@ -14,8 +14,7 @@
  */
 
 using System;
-using AWS.Lambda.PowerTools.Aspects;
-using AWS.Lambda.PowerTools.Core;
+using AWS.Lambda.PowerTools.Common;
 using AWS.Lambda.PowerTools.Logging.Internal;
 using Microsoft.Extensions.Logging;
 
@@ -23,9 +22,9 @@ namespace AWS.Lambda.PowerTools.Logging;
 
 /// <summary>
 ///     Class LoggingAttribute.
-///     Implements the <see cref="AWS.Lambda.PowerTools.Aspects.MethodAspectAttribute" />
+///     Implements the <see cref="MethodAspectAttribute" />
 /// </summary>
-/// <seealso cref="AWS.Lambda.PowerTools.Aspects.MethodAspectAttribute" />
+/// <seealso cref="MethodAspectAttribute" />
 [AttributeUsage(AttributeTargets.Method)]
 public class LoggingAttribute : MethodAspectAttribute
 {
