+++ libraries/src/AWS.Lambda.PowerTools.Tracing/AWS.Lambda.PowerTools.Tracing.csproj
@@ -20,7 +20,7 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <ProjectReference Include="..\AWS.Lambda.PowerTools\AWS.Lambda.PowerTools.csproj" />
+        <ProjectReference Include="..\AWS.Lambda.PowerTools.Common\AWS.Lambda.PowerTools.Common.csproj" />
     </ItemGroup>
 
     <ItemGroup>
@@ -28,7 +28,7 @@
         <PackageReference Include="AWSXRayRecorder.Core" Version="2.10.1" />
     </ItemGroup>
     <ItemGroup>
-        <None Include="README.md" Pack="true" PackagePath="\"/>
+        <None Include="README.md" Pack="true" PackagePath="\" />
     </ItemGroup>
     <ItemGroup>
         <None Include="../../AWSLogo128x128.png" Pack="true" Visible="false" PackagePath="" />
