+++ libraries/src/AWS.Lambda.PowerTools.Tracing/TracingCaptureMode.cs
@@ -13,7 +13,7 @@
  * permissions and limitations under the License.
  */
 
-namespace Amazon.Lambda.PowerTools.Tracing;
+namespace AWS.Lambda.PowerTools.Tracing;
 
 /// <summary>
 ///     Enum TracingCaptureMode
@@ -22,21 +22,21 @@ public enum TracingCaptureMode
 {
     /// <summary>
     ///     Enables attribute to capture only response. If this mode is explicitly overridden
-    ///     on {<see cref="T:Amazon.Lambda.PowerTools.Tracing.TracingAttribute" /> attribute, it will override value of
+    ///     on {<see cref="T:AWS.Lambda.PowerTools.Tracing.TracingAttribute" /> attribute, it will override value of
     ///     environment variable POWERTOOLS_TRACER_CAPTURE_RESPONSE
     /// </summary>
     Response,
 
     /// <summary>
     ///     Enabled attribute to capture only error from the method. If this mode is explicitly overridden
-    ///     on <see cref="T:Amazon.Lambda.PowerTools.Tracing.TracingAttribute" /> attribute, it will override value of
+    ///     on <see cref="T:AWS.Lambda.PowerTools.Tracing.TracingAttribute" /> attribute, it will override value of
     ///     environment variable POWERTOOLS_TRACER_CAPTURE_ERROR
     /// </summary>
     Error,
 
     /// <summary>
     ///     Enabled attribute to capture both response error from the method. If this mode is explicitly overridden
-    ///     on <see cref="T:Amazon.Lambda.PowerTools.Tracing.TracingAttribute" /> attribute, it will override value of
+    ///     on <see cref="T:AWS.Lambda.PowerTools.Tracing.TracingAttribute" /> attribute, it will override value of
     ///     environment variables POWERTOOLS_TRACER_CAPTURE_RESPONSE
     ///     and POWERTOOLS_TRACER_CAPTURE_ERROR
     /// </summary>
@@ -44,7 +44,7 @@ public enum TracingCaptureMode
 
     /// <summary>
     ///     Disables attribute to capture both response and error from the method. If this mode is explicitly overridden
-    ///     on <see cref="T:Amazon.Lambda.PowerTools.Tracing.TracingAttribute" /> attribute, it will override values of
+    ///     on <see cref="T:AWS.Lambda.PowerTools.Tracing.TracingAttribute" /> attribute, it will override values of
     ///     environment variable POWERTOOLS_TRACER_CAPTURE_RESPONSE
     ///     and POWERTOOLS_TRACER_CAPTURE_ERROR
     /// </summary>
