+++ libraries/tests/AWS.Lambda.PowerTools.Logging.Tests/LoggingAttributeTest.cs
@@ -19,8 +19,7 @@ using System.Linq;
 using Amazon.Lambda.APIGatewayEvents;
 using Amazon.Lambda.ApplicationLoadBalancerEvents;
 using Amazon.Lambda.Core;
-using AWS.Lambda.PowerTools.Aspects;
-using AWS.Lambda.PowerTools.Core;
+using AWS.Lambda.PowerTools.Common;
 using AWS.Lambda.PowerTools.Logging.Internal;
 using Microsoft.Extensions.Logging;
 using Moq;
